% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_repo.R
\name{add_papers}
\alias{add_papers}
\title{export papers from zotero to /relevant_papers of efficientize_exp_repo_template}
\usage{
add_papers(manuscript_file_name, exp_repo = ".")
}
\description{
export papers from zotero to /relevant_papers of efficientize_exp_repo_template
}
\examples{
add_papers('second_year_paper.Rmd', '/Users/xzfang/Github/talker_general_representation')
add_papers('version1.Rmd', '/Users/xzfang/Github/ding2016')

}
