% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_help.R
\name{capitalize_headings}
\alias{capitalize_headings}
\title{capitalize the first letter of the first word in rmd headings}
\usage{
capitalize_headings(script)
}
\arguments{
\item{script}{character}
}
\value{
character
}
\description{
capitalize the first letter of the first word in rmd headings
}
\examples{
capitalize_headings("### introduction")

}
