% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict_writing}
\alias{dict_writing}
\title{dictionary replace_in_selective_text_for_all_entries_in_all_dicts refers to}
\format{
dataframe
}
\usage{
dict_writing
}
\description{
replace shorthands with full spellings
}
\keyword{datasets}
