% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_help.R
\name{update_psychopy_script_for_multiple_lists}
\alias{update_psychopy_script_for_multiple_lists}
\title{apply functions capitalize_headings(), capitalize_sentences(), replace_based_on_dict() to an existing rmd}
\usage{
update_psychopy_script_for_multiple_lists(filename, output_filename = filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
apply functions capitalize_headings(), capitalize_sentences(), replace_based_on_dict() to an existing rmd
}
\examples{
update_psychopy_script_for_multiple_lists("/Users/xzfang/Github/ideal_adapter/interface_exp2.psyexp", "/Users/xzfang/Github/ideal_adapter/ideal_adapter_exp2_List_All/interface_exp2_list_all.psyexp")

}
