% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{capitalize_headings}
\alias{capitalize_headings}
\title{Title
capitalize first letter of every word in headings}
\usage{
capitalize_headings(script)
}
\arguments{
\item{script}{}
}
\value{

}
\description{
Title
capitalize first letter of every word in headings
}
\examples{
capitalize_headings("### introduction")
}
