% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{capitalize_sentences}
\alias{capitalize_sentences}
\title{capitalize the first letter of first word in each sentence}
\usage{
capitalize_sentences(script)
}
\arguments{
\item{script}{character}
}
\value{
character
}
\description{
capitalize the first letter of first word in each sentence
}
\examples{
capitalize_sentences("you are human. you really are. ")
}
