% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_help.R
\name{formalize_writing}
\alias{formalize_writing}
\title{apply functions capitalize_headings(), capitalize_sentences(), replace_in_selective_text_for_all_entries_in_all_dicts() to an existing rmd}
\usage{
formalize_writing(filename, output_filename = filename)
}
\arguments{
\item{filename}{}
}
\description{
apply functions capitalize_headings(), capitalize_sentences(), replace_in_selective_text_for_all_entries_in_all_dicts() to an existing rmd
}
