% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{formalize_writing}
\alias{formalize_writing}
\title{apply functions capitalize_headings() and capitalize_sentences() to an existing rmd}
\usage{
formalize_writing(filename, output_filename = filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
apply functions capitalize_headings() and capitalize_sentences() to an existing rmd
}
