% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_help.R
\name{replace_based_on_dict}
\alias{replace_based_on_dict}
\title{replace strings referring to "dictionaries"}
\usage{
replace_based_on_dict(script, dict_filenames = c("dict_writing.rda"))
}
\arguments{
\item{script}{character}
}
\value{
character
}
\description{
duh
}
\details{
i use it in two scenarios. for writing
,
replace shorthands with full spellings see https://github.com/xf15/efficientize/blob/main/data-raw/dict_writing.csv
.
for code
,
modify psyexp script used for a single list to be used as script for task combining all lists see https://github.com/xf15/efficientize/blob/main/data-raw/dict_psyexp.csv
.
if you want to use this feature, you should fork my repo, replace my csv with your own, and install your repo, don't send me pull request on this
}
\examples{
replace_based_on_dict("in exp1", 'dict_writing.rda')

}
